requirements.txt
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
cssselect2==0.8.0
defusedxml==0.7.1
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
pycparser==2.22
setuptools==80.9.0
sympy==1.14.0
tinycss2==1.4.0
typing_extensions==4.14.1
webencodings==0.5.1
tqdm==4.66.4
torch>=1.9.0
chess
numpy>=1.21.0
matplotlib>=3.3.4
Pillow>=8.3.1
python-chess